# [opengl] [perf] Use TI_AUTO_PROF in OpenGL backend runtime