<!--
Thanks for your PR!
If it's your first time contributing to Taichi, please make sure you have read our Contributor Guideline:
  https://taichi.readthedocs.io/en/latest/contributor_guide.html

- Please always prepend your PR title with tags such as [CUDA], [Lang], [Doc], [Example], e.g.:
    [Lang] Add ti.Complex as Taichi class
- Use a lowercased tag for PRs that are invisible to end-users, i.e., won't be highlighted in changelog:
    [cuda] [test] Fix out-of-memory error while running test
- More details: http://taichi.readthedocs.io/en/latest/contributor_guide.html#prtags

- Please fill the following blank with the issue number this PR related to (if any):
    Related issue = #2345
- If your PR will fix the issue **completely**, use the `close` or `fixes` keyword:
    Related issue = close #2345
- So that when the PR gets merged, GitHub will **automatically** close the issue #2345 for you.
- If the PR doesn't belong to any existing issue, and it's a trivial change, feel free to leave it blank :)
  -->
Related issue = #

[[Click here for the format server]](http://kun.csail.mit.edu:31415/)

----
```
          0.169 ms  0.03%  opengl_get_threads_per_group [18 x   9.378 us]
          0.137 ms  0.02%  compile               [7 x  19.550 us]
          3.498 ms  0.53%  link                  [7 x 499.657 us]
      6.088  s launch                        [25205 x 241.520 us]
          0.307  s  5.04%  dispatch_compute      [93605 x   3.277 us]
            171.665 ms 55.97%  glDispatchCompute     [93605 x   1.834 us]
             13.229 ms  4.31%  glMemoryBarrier       [93605 x 141.329 ns]
            121.825 ms 39.72%  [unaccounted]
          0.056  s  0.92%  launch:ext_arr1       [2400 x  23.377 us]
             55.077 ms 98.17%  add_buffer            [4800 x  11.474 us]
                 50.316 ms 91.36%  GLBuffer              [4800 x  10.483 us]
                     47.020 ms 93.45%  bind_data             [4800 x   9.796 us]
                      3.297 ms  6.55%  [unaccounted]
                  4.761 ms  8.64%  [unaccounted]
              1.027 ms  1.83%  [unaccounted]
          0.008  s  0.13%  add_buffer            [2400 x   3.377 us]
              7.083 ms 87.39%  GLBuffer              [2400 x   2.951 us]
                  6.223 ms 87.86%  bind_data             [2400 x   2.593 us]
                  0.860 ms 12.14%  [unaccounted]
              1.022 ms 12.61%  [unaccounted]
          5.686  s 93.40%  copy_back             [7200 x 789.696 us]
              5.665  s 99.64%  map                   [7200 x 786.832 us]
          0.031  s  0.51%  [unaccounted]
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
```
I found that currently `TI_AUTO_PROF` is always used in compile-time? But I thought it could also be useful in run-time performance test, is this intentional?