@archibate 2020-07-25
> Cool! Thanks.
> 
> Would it be possible to not introduce `benchmark_mpm99.py`? We already have too many copies in the repo.
> 
> > I found that currently TI_AUTO_PROF is always used in compile-time? But I thought it could also be useful in run-time performance test, is this intentional?
> 
> That's intentional. Kernel profiling has a very different need from compiler profiling.

So, we should need multiple `ScopedProfiler`?
1. compiler profiler
2. runtime profiler
3. python profiler (#1493)